Fixed a huge bug in StringBuilder bufferPrint().
